+++ KCS/KCS/Presentation/StoreUpdateRequest/ViewModel/Protocol/StoreUpdateRequestDependency.swift
@@ -1,0 +1,17 @@
+//
+//  StoreUpdateRequestDependency.swift
+//  KCS
+//
+//  Created by 김영현 on 2/20/24.
+//
+
+import RxSwift
+
+struct StoreUpdateRequestDepenency {
+    
+    let storeUpdateRequestUseCase: StoreUpdateRequestUseCase
+    let fetchStoreIDUseCase: FetchStoreIDUseCase
+    let setStoreIDUseCase: SetStoreIDUseCase
+    let disposeBag = DisposeBag()
+    
+}
